telemetry/requirements.txt
paho-mqtt
requests
requests
geoip2
